refactor : dto 패키지 사용 목적에 맞추어 정리
