+++ aiku/aiku-main/src/main/java/aiku_main/controller/LoginController.java
@@ -1,15 +1,13 @@
 package aiku_main.controller;
 
-import aiku_main.dto.RefreshTokenDto;
-import aiku_main.dto.RefreshTokenResDto;
-import aiku_main.dto.SignInDto;
-import aiku_main.dto.SignInTokenResDto;
+import aiku_main.dto.member.login.RefreshTokenDto;
+import aiku_main.dto.member.login.RefreshTokenResDto;
+import aiku_main.dto.member.login.SignInDto;
+import aiku_main.dto.member.login.SignInTokenResDto;
 import aiku_main.service.LoginService;
-import common.domain.member.Member;
 import common.response.BaseResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RequestMapping("/login")
