+++ aiku/aiku-main/src/main/java/aiku_main/controller/TermController.java
@@ -1,7 +1,7 @@
 package aiku_main.controller;
 
 import aiku_main.dto.DataResDto;
-import aiku_main.dto.TermResDto;
+import aiku_main.dto.member.TermResDto;
 import aiku_main.service.TermService;
 import common.domain.TermTitle;
 import common.response.BaseResponse;
@@ -10,8 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RequestMapping("/term/{termTitle}")
 @RequiredArgsConstructor
