+++ aiku/aiku-main/src/main/java/aiku_main/dto/member/AuthorityUpdateDto.java
@@ -1,15 +1,9 @@
-package aiku_main.dto;
+package aiku_main.dto.member;
 
-import common.domain.member.MemberProfileBackground;
-import common.domain.member.MemberProfileCharacter;
-import common.domain.member.MemberProfileType;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.web.multipart.MultipartFile;
 
 @Getter
 @NoArgsConstructor
