+++ aiku/aiku-main/src/main/java/aiku_main/dto/racing/ScheduleRacingMember.java
@@ -1,4 +1,4 @@
-package aiku_main.application_event.domain;
+package aiku_main.dto.racing;
 
 import aiku_main.dto.MemberProfileResDto;
 import common.domain.member.Member;
