+++ aiku/aiku-main/src/main/java/aiku_main/service/MemberFirebaseService.java
@@ -1,23 +1,12 @@
 package aiku_main.service;
 
 import aiku_main.dto.FirebaseTokenDto;
-import aiku_main.dto.TermResDto;
 import aiku_main.exception.*;
 import aiku_main.repository.MemberRepository;
-import aiku_main.repository.TermRepository;
-import common.domain.Term;
-import common.domain.TermTitle;
 import common.domain.member.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
-
-import static common.response.status.BaseErrorCode.DUPLICATED_FCM_TOKEN;
-import static common.response.status.BaseErrorCode.NO_FCM_TOKEN;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
