+++ aiku/aiku-main/src/main/java/aiku_main/service/MemberRegisterService.java
@@ -3,9 +3,9 @@ package aiku_main.service;
 import aiku_main.application_event.event.PointChangeReason;
 import aiku_main.application_event.event.PointChangeType;
 import aiku_main.application_event.publisher.PointChangeEventPublisher;
-import aiku_main.dto.MemberProfileDto;
-import aiku_main.dto.MemberRegisterDto;
-import aiku_main.dto.NicknameExistResDto;
+import aiku_main.dto.member.MemberProfileDto;
+import aiku_main.dto.member.MemberRegisterDto;
+import aiku_main.dto.member.NicknameExistResDto;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.oauth.KakaoOauthHelper;
 import aiku_main.repository.EventRepository;
@@ -14,7 +14,6 @@ import aiku_main.s3.S3ImageProvider;
 import common.domain.event.RecommendEvent;
 import common.domain.member.Member;
 import common.domain.member.MemberProfileType;
-import common.domain.value_reference.MemberValue;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
