+++ aiku/aiku-main/src/test/java/aiku_main/service/RacingServiceTest.java
@@ -1,8 +1,8 @@
 package aiku_main.service;
 
-import aiku_main.application_event.domain.ScheduleRacing;
-import aiku_main.application_event.domain.ScheduleRacingMember;
-import aiku_main.application_event.domain.ScheduleRacingResult;
+import aiku_main.dto.racing.ScheduleRacing;
+import aiku_main.dto.racing.ScheduleRacingMember;
+import aiku_main.dto.racing.ScheduleRacingResult;
 import aiku_main.repository.RacingQueryRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -15,7 +15,6 @@ import common.domain.member.MemberProfileType;
 import common.domain.schedule.Schedule;
 import common.domain.schedule.ScheduleMember;
 import common.domain.team.Team;
-import common.domain.value_reference.MemberValue;
 import common.domain.value_reference.TeamValue;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.AfterEach;
